[Qwen3] Shared state_dict_adapter class for MoE models
